debug: introduce replModel and move repl related methods inside it
