+++ src/vs/workbench/parts/debug/common/debug.ts
@@ -161,6 +161,7 @@ export interface IDebugSession extends ITreeElement {
 	removeReplExpressions(): void;
 	addReplExpression(stackFrame: IStackFrame, name: string): TPromise<void>;
 	appendToRepl(data: string | IExpression, severity: severity, source?: IReplElementSource): void;
+	logToRepl(sev: severity, args: any[], frame?: { uri: uri, line: number, column: number });
 
 	// session events
 	readonly onDidEndAdapter: Event<AdapterEndEvent>;
